--- tomcat/src/main/java/org/apache/coyote/http11/ErrorController.java
@@ -1,9 +1,0 @@
-package org.apache.coyote.http11;
-
-public class ErrorController implements Controller{
-
-    @Override
-    public HttpResponse handle(final HttpRequest request) {
-        return HttpResponse.toNotFound();
-    }
-}
