+++ tomcat/src/main/java/org/apache/coyote/http11/controller/ErrorController.java
@@ -1,0 +1,13 @@
+package org.apache.coyote.http11.controller;
+
+import org.apache.coyote.http11.Controller;
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
+
+public class ErrorController implements Controller {
+
+    @Override
+    public HttpResponse handle(final HttpRequest request) {
+        return HttpResponse.toNotFound();
+    }
+}
