+++ tomcat/src/main/java/org/apache/coyote/http11/controller/RegisterController.java
@@ -1,10 +1,16 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.controller;
 
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
+import org.apache.coyote.http11.ContentType;
+import org.apache.coyote.http11.Controller;
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.StatusCode;
+import org.apache.coyote.http11.ViewLoader;
 
-public class RegisterController implements Controller{
+public class RegisterController implements Controller {
 
     @Override
     public HttpResponse handle(final HttpRequest request) {
