+++ tomcat/src/main/java/org/apache/coyote/http11/controller/StaticController.java
@@ -1,4 +1,10 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.controller;
+
+import org.apache.coyote.http11.Controller;
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.StatusCode;
+import org.apache.coyote.http11.ViewLoader;
 
 public class StaticController implements Controller {
